# lib/api-zod/src/generated/types/applicationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PhD Application Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface ApplicationInput {
  /** @minLength 1 */
  university: string;
  /** @minLength 1 */
  country: string;
  program?: string;
  department?: string;
  projectTitle?: string;
  supervisorName?: string;
  fundingSource?: string;
  portalLink?: string;
  fieldArea?: string;
  stage?: string;
  dateSourced?: string;
  deadline?: string;
  dateApplied?: string;
  interviewDate?: string;
  outcomeDate?: string;
  offerStatus?: string;
  notes?: string;
  priorityScore?: number;
  fitScore?: number;
  statusReason?: string;
  nextAction?: string;
  nextActionDate?: string;
  researchFitReason?: string;
  supervisorFitReason?: string;
  papersToRead?: string;
  skillsMatch?: string;
  missingRequirements?: string;
  risks?: string;
  outreachAngle?: string;
}