# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PhD Application Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: number;
  applicationId: number;
  docType: string;
  isRequired?: boolean;
  isUploaded?: boolean;
  isVerified?: boolean;
  isSent?: boolean;
  /** @nullable */
  sentDate?: string | null;
  /** @nullable */
  versionNumber?: number | null;
  /** @nullable */
  fileName?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}