# lib/api-zod/src/generated/types/outreachUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PhD Application Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface OutreachUpdate {
  applicationId?: number;
  professorName?: string;
  institution?: string;
  email?: string;
  dateContacted?: string;
  subjectLine?: string;
  messageSummary?: string;
  threadStatus?: string;
  followUpCount?: number;
  lastFollowUpDate?: string;
  nextFollowUpDate?: string;
  responseReceived?: boolean;
  responseSummary?: string;
  outcome?: string;
  toneQuality?: string;
  attachmentsSent?: string;
}